% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dastash-package.R
\docType{package}
\name{dastash-package}
\alias{dastash}
\alias{dastash-package}
\title{dastash: Persistent Disk Cache Shared Between Processes}
\description{
A persistent, size-bounded cache that several 'R' processes on one machine can share, in the style of the 'Python' library 'diskcache'. Keys map to values that outlive the session, expire on a clock, carry tags for bulk invalidation, and are evicted when the cache grows past a limit. Metadata and small values live in a transactional 'libmdbx' database through the 'mdbx' package, and larger values are stored as content-addressed files. Functions can be memoised across sessions and processes, and the cache can be used through the 'cachem' interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pedrobtz.github.io/dastash/}
  \item \url{https://github.com/pedrobtz/dastash}
  \item Report bugs at \url{https://github.com/pedrobtz/dastash/issues}
}

}
\author{
\strong{Maintainer}: Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Pedro Baltazar \email{pedrobtz@gmail.com} [copyright holder]
}

}
\keyword{internal}
